restorant_proj/requirements.txt
python-dotenv==0.14.0
psycopg2-binary==2.8.5

Django==3.0.9
djangorestframework==3.11.0

# Testing
pytest==6.0.1
pytest-django==3.9.0
flake8==3.8.3
flake8-django==1.1.1
factory_boy==2.12.0

# Production
gunicorn==20.0.4

# Development
django_extensions==3.0.3